small refacto: move some tests file
now all tests that need a mongo db needs to be in integration/mongo
